requirements.txt
fastapi
uvicorn
pytest
httpx
PyYAML
